[#56] consolidate progress bar logic in mixin
