+++ src/components/mixins/bestPracticeData.js
@@ -34,6 +34,15 @@ export const bestPracticeData = {
           query.where('type', activityType)
         })
         .count()
+    },
+    percentBPActivitesByType: function (activityType) {
+      const activitiesWithBP = this.$store.getters['entities/activities/query']().whereHas('assessments', (query) => {
+        query.where('value', [this.$t('bestPracticeOptions.yesKey')])
+      }).where('type', activityType).count()
+
+      const activitiesInType = this.$store.getters['entities/activities/query']().where('type', activityType).count()
+
+      return (activitiesWithBP / activitiesInType).toFixed(2) * 100
     }
   }
 }
